chore: move function `get_package_dir`into `packsmanager.py`
