+++ news/move-get-packages-dir.rst
@@ -1,0 +1,23 @@
+**Added:**
+
+* No news added: Move function get_package_dir from __init__.py into packsmanager.py.
+
+**Changed:**
+
+* <news item>
+
+**Deprecated:**
+
+* <news item>
+
+**Removed:**
+
+* <news item>
+
+**Fixed:**
+
+* <news item>
+
+**Security:**
+
+* <news item>
